+++ client/src/test/e2e.test.ts
@@ -1,10 +1,16 @@
-import { ErrorAction, fixturesFolder, GetDefinitionAction, GetUsagesAction, readUseCases, UseCase, UseCaseAnnotation } from "./useCasesHelper";
-import { constructRange, getDocUri, getErrors, openUseCaseFile } from './vscodeHelper';
+import { ErrorAction, fixturesFolder, GetDefinitionAction, GetUsagesAction, HoveringAction, readUseCases, UseCase, UseCaseAnnotation } from "./useCasesHelper";
+import { constructRange, flushCodeCoverage, getDocUri, getErrors, openUseCaseFile } from './vscodeHelper';
 import * as assert from 'assert';
 import * as path from 'path';
-import { Range, DiagnosticSeverity, commands, Location } from "vscode";
+import { Range, DiagnosticSeverity, commands, Location, Hover, MarkdownString } from "vscode";
 
 const useCases = readUseCases();
+
+export const mochaHooks = {
+    async afterAll() {
+      await flushCodeCoverage();
+    }
+};
 
 suite('Description is valid for:', () => {
     for (const useCase of useCases) {
@@ -25,6 +31,9 @@ suite('Description is valid for:', () => {
                     case "DefinitionResult":
                     case "UsagesResult":
                         // Nothing to check here
+                        break;
+                    case 'Hovering':
+                        assert.ok('text' in action, `Action ${annotation.name} has no text`);
                         break;
                     default:
                         assert.fail(`Unknown action ${JSON.stringify(action)}`);
@@ -93,6 +102,25 @@ for (const useCase of useCases) {
             }
         });
 
+        test("check hovering", async () => {
+            for (const getHoveringAnnotation of annotations.filter(it => it.action.type === 'Hovering')) {
+                const text = (getHoveringAnnotation.action as HoveringAction).text;
+                const notMatch = (getHoveringAnnotation.action as HoveringAction).not || false;
+                const hover = await commands.executeCommand<Hover[]>('vscode.executeHoverProvider', mainUri, {
+                    line: getHoveringAnnotation.range.line,
+                    character: getHoveringAnnotation.range.startChar,
+                });
+                if (hover.length == 0) {
+                    if (notMatch) continue;
+                    assert.fail(`Hovering is expected here, but there is no any.\n ${printLineWithRange(useCase, getRange(getHoveringAnnotation))}`);
+                }
+                const message = (hover[0].contents[0] as MarkdownString).value;
+                const matched = message.includes(text);
+                if (matched == notMatch)
+                    assert.fail(`Hovering has something not expected.\n ${printLineWithRange(useCase, getRange(getHoveringAnnotation))}\n Expected: ${text}, but was: ${message}`);
+            }
+        });
+
         // all vscode commands here https://code.visualstudio.com/api/references/commands
     });
 }
@@ -123,16 +151,20 @@ suite('Test include', () => {
         assert.deepEqual(anotherDefinition.range, constructRange(0, 0, 7));
     });
 
-    test('Errors dissapear on closing', async () => {
+    // There is a bug in VSCode with cleaning diagnostics
+    test.skip('Errors dissapear on closing', async () => {
         const mainUri = getDocUri(path.resolve(includeFolder, 'withError.asm'));
         await openUseCaseFile(mainUri);
-        await getErrors(mainUri, 0);
+        let errors = await getErrors(mainUri, 0);
+        assert.equal(errors.length, 0);
         const includeUri = getDocUri(path.resolve(includeFolder, 'errorInclude.asm'));
         const includeError = (await getErrors(includeUri, 1))[0];
         assert.deepEqual(includeError.message, 'Label is not defined');
         await commands.executeCommand('workbench.action.closeActiveEditor');
-        await getErrors(mainUri, 0);
-        await getErrors(includeUri, 0);
+        errors = await getErrors(mainUri, 0);
+        assert.equal(errors.length, 0);
+        errors = await getErrors(includeUri, 0);
+        assert.equal(errors.length, 0);
     });
 });