+++ server/src/parser/ast/macros.ts
@@ -1,10 +1,10 @@
 import { UnifiedCommandName } from "../../validators/asmCommandValidator";
-import { BasicNode } from "./nodes";
+import { AstNode } from "./nodes";
 
 export type MacrosByName = Map<UnifiedCommandName, MacrosObject>
 
 export type MacrosObject = {
     name: UnifiedCommandName,
-    definitions: BasicNode[],
-    usages: BasicNode[],
+    definitions: AstNode[],
+    usages: AstNode[],
 }