+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Added
+- hovering on asm mnemonics
+
 ## [1.0.0] - 2022-09-09
 ### Added
 - extension description